distinct counter interface refactoring
